fix(testing): keep shared gate proofs code-only
